+++ scout/parse/variant/variant.py
@@ -2,7 +2,7 @@ import logging
 
 from scout.constants import CHR_PATTERN
 from scout.exceptions import VcfError
-from scout.utils import call_safe
+from scout.utils.convert import call_safe
 from scout.utils.dict_utils import remove_nonetype
 
 from .callers import parse_callers
